ResumeAnalyzerBasedOnAI/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.4",
    "langchain>=0.3.18",
    "pypdf2>=3.0.1",
    "python-multipart>=0.0.20",
]
